Updated code as per modified sp

+++ IncidentAPI/Dtos/Responses/NameAndIncidentCountByPriorityResponse.cs
@@ -5,6 +5,6 @@ namespace Incident.API.Dtos.Responses
         public string AssignedToName { get; set; } = string.Empty;
         public string Priority { get; set; } = string.Empty;
         public int IncidentCount { get; set; }
-        public int? AvgResolutionTime_Hours { get; set; }
+        public double? AvgResolutionTime_Hours { get; set; }
     }
 }
